Interactive Calculator
Built with Racket.

All 6 operations are supported. Users can denote the use of negative
number by using brackets and the minus sign – negative 6 can be written
as (-6). Errors are handled with try-catch-like structure that prints
out the error instead of crashing the machine – which synergizes well
with my idea of throwing an error whenever the user enters something
that doesn’t make sense. Data protection was not a priority, which
explains the definitions of a lot of global variables. I/O is supported
– input is for evaluating a batch of equations in a file, and output
prints out the calculation into a text file of user’s choice.